requirements.txt
arabic-reshaper==3.0.0
asgiref==3.7.2
asn1crypto==1.5.1
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.1
cssselect2==0.7.0
Django==5.0.1
django-ckeditor==6.7.0
django-js-asset==2.2.0
html5lib==1.1
idna==3.6
lxml==5.1.0
oscrypto==1.3.0
pillow==10.2.0
pycparser==2.21
pyHanko==0.21.0
pyhanko-certvalidator==0.26.3
pypdf==4.0.1
pypng==0.20220715.0
python-bidi==0.4.2
PyYAML==6.0.1
qrcode==7.4.2
reportlab==4.0.9
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
svglib==1.5.1
tinycss2==1.2.1
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
uritools==4.0.2
urllib3==2.1.0
webencodings==0.5.1
xhtml2pdf==0.2.14
